many: add initial implementation of etrace
strace.go is a forked implementation of the strace package from snapd. It has
been removed so we do not depend directly on snapd and we can build this without
needing to land code in snapd for this to work. Some improvements made here will
eventually be landed back to upstream snapd.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>