strace.go: don't divide by time.Nanosecond
This is effectively a no-op and could confuse the reader into thinking something
else is going on, when all we really want to do is turn the float of the time in
seconds into a time.Duration stored as a float.

Fixes regression introduced by f8155823dff88357f3d2a6db32e5bdad98dd6d65.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>