main.go: add --output-file option to write just the output to a file
This makes it easier to measure programs which are finicky and want to crash the
terminal window if they can't write to stdout for some reason.

Also open the log files in append mode, so that the log files show all of the
output from iterations, previously we would just re-write from the beginning of
the file each iteration.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>