many: implement file tracing
This adds a program to trace the files used during execution by parsing the
strace output with various regular expressions.

This will catch almost all file accesses, with the following caveats:
- It will not handle properly syscalls which have multiple path arguments if the
  second path matches, it will only catch the first path.
- It will not handle syscalls with a fd and a relative path if the fd is
  AT_FDCWD since we don't currently have a reliable way to track the current
  working directory for a process and strace doesn't easily give us that
  information

Note that there is still a fair amount of overlap between this code and the main
etrace code as well as between the TraceExecveWithFiles and TraceExecveTimings
function.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>