internal/strace/file-tracing.go: refactor file regexes
Refactor to apply the file regexes at the end, and hash the file
accesses not just by the path but also by the program that accessed the
file and the pid so that we don't keep duplicate paths for the same
program.

This will allow us to output the programs that access files and also
will allow us to eventually filter on what programs should have their
files tracked/reported.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>