cmd/etrace: support filtering files by program name
Program name filtering can be done either with a regex to match the
program name, or by a black list of patterns. Ideally we would do
everything with a pattern but there may be cases where the blacklist
feature is desirable in combination with the regex filtering because the
prototypical use case for the black list is to filter out snapd binaries
from the output. Trying to accomplish both of these goals with just a
single regex is just too hard to do considering that go stdlib regex
doesn't support negative lookaheads.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>