cmd/etrace: stop falling back to using wmctrl
wmctrl with the current args we are providing has a habit of matching other
unrelated windows, in some case the window name ends up being "", in which case
this will close a random application that is open at the time.

Since we already added code to kill the process for a window ID if closing via
xdotool fails, and we have additional options to close processes there using
cgroups that snap apps are put into now, just drop support for using wmctrl.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>