cmd/etrace/cmd_file.go: remove unnecessary output/measuring for file command
The file command is meant for determining info about what files were accessed,
not how long it takes to start up, so drop the info about how long it took to
display a window from the default output. Still include it in the JSON output
though in case folks are consuming it there.

Also output the non-json version after running the restore script.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>