cmd/etrace: fix error serialization in JSON
Errors were being serialized as "{}" since the error type is technically
treated as a struct, but with no exported fields. Instead the type we want to
use to serialize errors is a string, which allows saving the error message.

Still to do is to handle errors from the sub-process etrace that is called in
analyze-snap, but for now as long as there is a valid number for each of the
measurements we still treat it as valid.

Also fix a typo.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>